+++ src/main/java/ru/yandex/practicum/filmorate/service/DirectorService.java
@@ -19,7 +19,7 @@ public class DirectorService {
         return directorRepository.findAll();
     }
 
-    public Director findById(Long id) {
+    public Director findById(Integer id) {
         return directorRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Director not found"));
     }
@@ -32,7 +32,7 @@ public class DirectorService {
         return directorRepository.update(newDirector);
     }
 
-    public void delete(Long id) {
+    public void delete(Integer id) {
         directorRepository.delete(id);
     }
 }
